:sparkles: Add Place controller

+++ src/database/migrations/2018_01_09_232125_create_place_table.php
@@ -21,8 +21,8 @@ class CreatePlaceTable extends Migration
             $table->string('zip_code');
             $table->string('city');
             $table->string('country');
-            $table->decimal('latitude');
-            $table->decimal('longitude');
+            $table->float('latitude', 10, 6);
+            $table->float('longitude', 10, 6);
 
             $table->timestamps();
         });
